refactor: Introduce ProductContext for global state management
Set up a ProductContext to manage the state of product data globally
